requirements.txt
flask
flask-cors
ultralytics
easyocr
opencv-python
pillow
numpy